+++ src/forge_cli/sdk/config.py
@@ -1,0 +1,3 @@
+import os
+
+BASE_URL = os.environ.get("KNOWLEDGE_FORGE_URL", "http://localhost:9999")
